Change Name to Identifier, use different graphql lib
Graphql-go/graphql has proven to be overcomplicated and a giant pain in
the butt. neelance/graphql-go, on the other hand, looks simple and
idiomatic.

Name on all structs has been changed to Identifier. This was because it
was conflicting with a function name for graphql-go/graphql. Now that I'm
not using that library, it is too much of a pain to change the names,
meh.