Redo queries for new database
https://github.com/veekun/pokedex already has an excellent database of
pokemon info, and provides an easy way to import it into your database
of choice. Thus, all models will be changed to work with the new
database schema. All models that work for the old database have been
deleted for now, and will be rewritten at a later time. The database
name has also been changed to pokedex