Add comments, interfaces, and updates
Created interfaces for finding data for a generation and a region. The
Finder interface will be a composition of all interfaces like these.
This makes DB a Finder. A model also now has a retriver field that says
how to retrive models related to it by some key. These are composed of
the finders that a model needs to find its related info.

The info string for the sqlx.Open has been changed because the previous
non url version failed to work for some reason, i dont know why.

DB now has a field called session that is the sqlx.DB connection, thus
hiding all sqlx.DB methods from the public DB